+++ lib/ModelMixins/MappableMixin.ts
@@ -3,6 +3,7 @@ import { computed, runInAction } from "mobx";
 import Rectangle from "terriajs-cesium/Source/Core/Rectangle";
 import TerrainProvider from "terriajs-cesium/Source/Core/TerrainProvider";
 import DataSource from "terriajs-cesium/Source/DataSources/DataSource";
+import Cesium3DTileset from "terriajs-cesium/Source/Scene/Cesium3DTileset";
 import ImageryProvider from "terriajs-cesium/Source/Scene/ImageryProvider";
 import AsyncLoader from "../Core/AsyncLoader";
 import Constructor from "../Core/Constructor";
@@ -41,6 +42,12 @@ export namespace ImageryParts {
 
 export function isPrimitive(mapItem: MapItem): mapItem is AbstractPrimitive {
   return "isDestroyed" in mapItem;
+}
+
+export function isCesium3DTileset(
+  mapItem: MapItem
+): mapItem is Cesium3DTileset {
+  return "allTilesLoaded" in mapItem;
 }
 
 export function isTerrainProvider(